client/agent/requirements.txt
requests==2.31.0
cryptography==41.0.7
python-dotenv==1.0.0
colorama==0.4.6 